chore(config): split config into a pkg
so that the admission controller can be implemented separately from the
controller manager
