+++ railgun/cmd/rootcmd/rootcmd.go
@@ -7,17 +7,18 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/kong/kubernetes-ingress-controller/railgun/manager"
+	"github.com/kong/kubernetes-ingress-controller/railgun/pkg/config"
 )
 
-var config manager.Config
+var cfg config.Config
 
 func init() {
-	rootCmd.Flags().AddFlagSet(config.FlagSet())
+	rootCmd.Flags().AddFlagSet(cfg.FlagSet())
 }
 
 var rootCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
-		return manager.Run(cmd.Context(), &config)
+		return manager.Run(cmd.Context(), &cfg)
 	},
 	SilenceUsage: true,
 }
